Parse stderr for unadf banner.
Unadf does return 0 exit code always, even, if there was an issue in
either parsing image, mounting it or on any other issue. To be worst, it
always throw an banner to the stderr, so it cannot be used to check if
there was an error either.

As a workaround, let's check if there is a banner, strip it out and then
see if anything extra is on the stderr, so that it can be used to see if
there was any issues.